Release v1.2
Split goals into uploadAppPackage, uploadTestPackage, uploadExtraData and run.
Add documentation.
